requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
# Поддержка AVI и MP4 форматов 
# Progress bar library
tqdm==4.66.4
# SSIM calculation
scikit-image==0.21.0
# Data analysis and CSV export
pandas==2.0.3 